requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
Werkzeug==2.3.7
geopy==2.4.0
uszipcode==1.0.1
pyzipcode==0.2.1
gunicorn==21.2.0
